+++ src/main/java/com/plms/springframework/bean/factory/BeanFactory.java
@@ -22,4 +22,13 @@ public interface BeanFactory {
      * @return bean实例
      */
     Object getBean(String beanName, Object... args);
+
+    /**
+     * 根据类型获取bean实例
+     * @param beanName bean名称
+     * @param requiredType 要求的类型
+     * @param <T> 要求的类型
+     * @return bean实例
+     */
+    <T> T getBean(String beanName, Class<T> requiredType);
 }