+++ src/main/java/com/plms/springframework/bean/factory/HierarchicalBeanFactory.java
@@ -1,0 +1,8 @@
+package com.plms.springframework.bean.factory;
+
+/**
+ * @Author bigboss
+ * @Date 2021/11/3 13:17
+ */
+public interface HierarchicalBeanFactory extends BeanFactory{
+}