+++ src/main/java/com/plms/springframework/bean/factory/config/AutowireCapableBeanFactory.java
@@ -1,0 +1,10 @@
+package com.plms.springframework.bean.factory.config;
+
+import com.plms.springframework.bean.factory.BeanFactory;
+
+/**
+ * @Author bigboss
+ * @Date 2021/11/3 13:16
+ */
+public interface AutowireCapableBeanFactory extends BeanFactory {
+}