+++ src/main/java/com/plms/springframework/bean/factory/config/ConfigurableBeanFactory.java
@@ -1,0 +1,11 @@
+package com.plms.springframework.bean.factory.config;
+
+import com.plms.springframework.bean.factory.HierarchicalBeanFactory;
+
+/**
+ * @Author bigboss
+ * @Date 2021/11/3 13:16
+ */
+public interface ConfigurableBeanFactory extends HierarchicalBeanFactory, SingletonBeanRegistry {
+
+}