+++ src/main/java/com/plms/springframework/bean/factory/support/AbstractBeanDefinitionReader.java
@@ -1,0 +1,34 @@
+package com.plms.springframework.bean.factory.support;
+
+import com.plms.springframework.core.io.DefaultResourceLoader;
+import com.plms.springframework.core.io.ResourceLoader;
+
+/**
+ * @Author bigboss
+ * @Date 2021/11/3 13:59
+ */
+public abstract class AbstractBeanDefinitionReader implements BeanDefinitionReader{
+
+    private final BeanDefinitionRegistry registry;
+
+    private ResourceLoader resourceLoader;
+
+    public AbstractBeanDefinitionReader(BeanDefinitionRegistry registry) {
+        this(registry, new DefaultResourceLoader());
+    }
+
+    public AbstractBeanDefinitionReader(BeanDefinitionRegistry registry, ResourceLoader resourceLoader) {
+        this.registry = registry;
+        this.resourceLoader = resourceLoader;
+    }
+
+    @Override
+    public BeanDefinitionRegistry getRegistry() {
+        return registry;
+    }
+
+    @Override
+    public ResourceLoader getResourceLoader() {
+        return resourceLoader;
+    }
+}