+++ src/main/java/com/plms/springframework/bean/factory/support/BeanDefinitionRegistry.java
@@ -13,4 +13,25 @@ public interface BeanDefinitionRegistry {
      * @param beanDefinition
      */
     void registerBeanDefinition(String beanName, BeanDefinition beanDefinition);
+
+    /**
+     * 根据bean名称获取定义
+     * @param beanName
+     * @return
+     */
+    BeanDefinition getBeanDefinition(String beanName);
+
+    /**
+     * 获取所有Bean的名称
+     * @return
+     */
+    String[] getBeanDefinitionNames();
+
+    /**
+     * 是否包含指定名称的BeanDefinition
+     *
+     * @param beanName
+     * @return
+     */
+    boolean containsBeanDefinition(String beanName);
 }