+++ src/main/java/com/plms/springframework/bean/factory/xml/XmlBeanDefinitionReader.java
@@ -1,0 +1,111 @@
+package com.plms.springframework.bean.factory.xml;
+
+import cn.hutool.core.util.StrUtil;
+import cn.hutool.core.util.XmlUtil;
+import com.plms.springframework.bean.BeansException;
+import com.plms.springframework.bean.PropertyValue;
+import com.plms.springframework.bean.factory.config.BeanDefinition;
+import com.plms.springframework.bean.factory.config.BeanReference;
+import com.plms.springframework.bean.factory.support.AbstractBeanDefinitionReader;
+import com.plms.springframework.bean.factory.support.BeanDefinitionRegistry;
+import com.plms.springframework.core.io.Resource;
+import com.plms.springframework.core.io.ResourceLoader;
+import org.w3c.dom.Document;
+import org.w3c.dom.Element;
+import org.w3c.dom.Node;
+import org.w3c.dom.NodeList;
+
+import java.io.IOException;
+import java.io.InputStream;
+
+/**
+ * @Author bigboss
+ * @Date 2021/11/3 13:17
+ */
+public class XmlBeanDefinitionReader extends AbstractBeanDefinitionReader {
+
+
+    public XmlBeanDefinitionReader(BeanDefinitionRegistry registry) {
+        super(registry);
+    }
+
+    public XmlBeanDefinitionReader(BeanDefinitionRegistry registry, ResourceLoader resourceLoader) {
+        super(registry, resourceLoader);
+    }
+
+    @Override
+    public void loadBeanDefinitions(Resource resource) throws BeansException {
+        try(InputStream inputStream = resource.getInputStream()) {
+            doLoadBeanDefinitions(inputStream);
+        } catch (IOException e) {
+            throw new BeansException("IOException parsing XML document from " + resource, e);
+        }
+    }
+
+    @Override
+    public void loadBeanDefinitions(Resource... resources) throws BeansException {
+        for (Resource resource : resources) {
+            loadBeanDefinitions(resource);
+        }
+    }
+
+    @Override
+    public void loadBeanDefinitions(String location) throws BeansException {
+        Resource resource = getResourceLoader().getResource(location);
+        loadBeanDefinitions(resource);
+    }
+
+    protected void doLoadBeanDefinitions(InputStream inputStream) {
+        Document document = XmlUtil.readXML(inputStream);
+        Element root = document.getDocumentElement();
+        NodeList childNodes = root.getChildNodes();
+        for (int i = 0; i < childNodes.getLength(); i++) {
+            if (!(childNodes.item(i) instanceof Element)) {
+                continue;
+            }
+            if (!"bean".equals(childNodes.item(i).getNodeName())) {
+                continue;
+            }
+            Element bean = (Element) childNodes.item(i);
+            String id = bean.getAttribute("id");
+            String name = bean.getAttribute("name");
+            String beanClassName = bean.getAttribute("class");
+            Class<?> beanClass = null;
+            try {
+                beanClass = Class.forName(beanClassName);
+            } catch (ClassNotFoundException e) {
+                throw new BeansException("can not find class named [" + beanClassName + "]");
+            }
+            String beanName = StrUtil.isNotEmpty(id) ? id : name;
+            if (StrUtil.isEmpty(beanName)) {
+                beanName = StrUtil.lowerFirst(beanClass.getSimpleName());
+            }
+            BeanDefinition beanDefinition = new BeanDefinition(beanClass);
+            for (int j = 0; j < bean.getChildNodes().getLength(); j++) {
+                if (!(childNodes.item(i) instanceof Element)) {
+                    continue;
+                }
+                if (!"property".equals(bean.getChildNodes().item(j).getNodeName())) {
+                    continue;
+                }
+                Element property = (Element) bean.getChildNodes().item(j);
+                String attrName = property.getAttribute("name");
+                String attrValue = property.getAttribute("value");
+                String attrRef = property.getAttribute("ref");
+                if (StrUtil.isEmpty(attrName)) {
+                    throw new BeansException("the name of bean can not be empty!");
+                }
+                Object value = attrValue;
+                if (StrUtil.isNotEmpty(attrRef)) {
+                    value = new BeanReference(attrRef);
+                }
+                PropertyValue propertyValue = new PropertyValue(attrName, value);
+                beanDefinition.getPropertyValues().addPropertyValue(propertyValue);
+            }
+            if (getRegistry().containsBeanDefinition(beanName)) {
+                throw new BeansException("BeanName [" + beanName +"] is duplicate");
+            }
+            getRegistry().registerBeanDefinition(beanName, beanDefinition);
+        }
+    }
+}