+++ src/test/java/com/plms/springframework/BeanInstantiationStrategyTest.java
@@ -1,6 +1,5 @@
-package com.plms.springframework.bean;
+package com.plms.springframework;
 
-import cn.hutool.core.lang.Assert;
 import com.plms.springframework.bean.factory.config.BeanDefinition;
 import com.plms.springframework.bean.factory.support.DefaultListableBeanFactory;
 import com.plms.springframework.service.HelloService;
@@ -8,16 +7,16 @@ import org.junit.Test;
 
 /**
  * @Author bigboss
- * @Date 2021/11/1 22:20
+ * @Date 2021/11/2 14:37
  */
-public class BeanDefinitionAndBeanRegistyrTest {
+public class BeanInstantiationStrategyTest {
 
     @Test
     public void test_BeanFactory() {
         DefaultListableBeanFactory factory = new DefaultListableBeanFactory();
         BeanDefinition beanDefinition = new BeanDefinition(HelloService.class);
         factory.registerBeanDefinition("helloService", beanDefinition);
-        HelloService helloService = (HelloService) factory.getBean("helloService");
-        Assert.isTrue(helloService.hello().equals("hello world"));
+        HelloService bean = (HelloService) factory.getBean("helloService", "promise");
+        System.out.println(bean.hello());
     }
 }