+++ src/test/java/com/plms/springframework/bean/Car.java
@@ -1,0 +1,13 @@
+package com.plms.springframework.bean;
+
+import lombok.Data;
+
+/**
+ * @Author bigboss
+ * @Date 2021/11/3 14:39
+ */
+@Data
+public class Car {
+
+    private String brand;
+}